requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
PyMySQL
python-dotenv
firebase-admin
pandas
python-multipart